LayoutTests/fast/js/mozilla/strict/15.4.4.8-expected.txt

PASS 'use strict'; var a = arr(); a.reverse() threw exception of type TypeError.
PASS var a = arr(); a.reverse() threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = obj(); Array.prototype.reverse.call(o) threw exception of type TypeError.
PASS var o = obj(); Array.prototype.reverse.call(o) threw exception of type TypeError.
PASS true === true
FAIL 'use strict'; var a = agap(); a.reverse() should throw an instance of TypeError
FAIL var a = agap(); a.reverse() should throw an instance of TypeError
PASS true === true
FAIL 'use strict'; var o = ogap(); Array.prototype.reverse.call(o) should throw an instance of TypeError
FAIL var o = ogap(); Array.prototype.reverse.call(o) should throw an instance of TypeError
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
